Before action for login

+++ db/migrate/20170719143659_create_beers.rb
@@ -5,7 +5,7 @@ class CreateBeers < ActiveRecord::Migration[5.1]
       t.integer :user_id
       t.string :style
       t.float :abv
-      t.float :ibu
+      t.integer :ibu
       t.float :srm
 
       t.timestamps
